migrated opcodes air out of cairo air
