+++ stwo_cairo_prover/crates/prover/src/cairo_air/mod.rs
@@ -1,4 +1,5 @@
 pub mod air;
+pub mod opcodes_air;
 
 use air::{lookup_sum, CairoClaimGenerator, CairoComponents, CairoInteractionElements, CairoProof};
 use num_traits::Zero;
@@ -17,7 +18,7 @@ use crate::input::CairoInput;
 
 const LOG_MAX_ROWS: u32 = 20;
 
-const IS_FIRST_LOG_SIZES: [u32; 7] = [18, 4, 14, 19, 7, 6, 5];
+const IS_FIRST_LOG_SIZES: [u32; 16] = [19, 18, 17, 16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4];
 pub fn prove_cairo(input: CairoInput) -> Result<CairoProof<Blake2sMerkleHasher>, ProvingError> {
     let _span = span!(Level::INFO, "prove_cairo").entered();
     let config = PcsConfig::default();
